Large ogg file support
Similar to the previous patch, this patch removes two kinds of limitions:
 - 32-bit filesize
 - 32-bit frame and sample numbers
The latter is likely not a limitation we'll ever hit, but because libogg does
support 64-bit frame/sample numbers I decided to not truncate those number to
32-bit in our OggReader class.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13089 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/laserdisc/LaserdiscPlayer.hh
@@ -94,7 +94,7 @@ private:
 	void pause(EmuTime::param time);
 	void stop(EmuTime::param time);
 	void eject(EmuTime::param time);
-	void seekFrame(int frame, EmuTime::param time);
+	void seekFrame(size_t frame, EmuTime::param time);
 	void stepFrame(bool);
 	void seekChapter(int chapter, EmuTime::param time);
 
@@ -135,7 +135,7 @@ private:
 
 	void nextFrame(EmuTime::param time);
 	void setFrameStep();
-	int currentFrame;
+	size_t currentFrame;
 	int frameStep;
 
 	// Audio state
@@ -168,7 +168,7 @@ private:
 	SeekState seekState;
 
 	/* frame the MSX has requested to wait for */
-	int waitFrame;
+	size_t waitFrame;
 
 	// pause playing back on reaching wait frame
 	bool stillOnWaitFrame;
